+++ src/main/java/in/koreatech/koin/domain/user/service/UserService.java
@@ -7,6 +7,7 @@ import java.util.UUID;
 
 import in.koreatech.koin.domain.student.model.redis.StudentTemporaryStatus;
 import in.koreatech.koin.domain.student.repository.StudentRedisRepository;
+
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
@@ -16,14 +17,10 @@ import in.koreatech.koin.domain.owner.repository.OwnerRepository;
 import in.koreatech.koin.domain.timetableV2.repository.TimetableFrameRepositoryV2;
 import in.koreatech.koin.domain.user.dto.AuthResponse;
 import in.koreatech.koin.domain.user.dto.CoopResponse;
-import in.koreatech.koin.domain.user.dto.EmailCheckExistsRequest;
-import in.koreatech.koin.domain.user.dto.NicknameCheckExistsRequest;
 import in.koreatech.koin.domain.user.dto.UserLoginRequest;
 import in.koreatech.koin.domain.user.dto.UserLoginResponse;
-import in.koreatech.koin.domain.user.dto.UserPasswordCheckRequest;
 import in.koreatech.koin.domain.user.dto.UserTokenRefreshRequest;
 import in.koreatech.koin.domain.user.dto.UserTokenRefreshResponse;
-import in.koreatech.koin.domain.user.exception.DuplicationNicknameException;
 import in.koreatech.koin.domain.user.model.User;
 import in.koreatech.koin.domain.user.model.UserDeleteEvent;
 import in.koreatech.koin.domain.user.model.UserToken;
@@ -33,7 +30,6 @@ import in.koreatech.koin.domain.user.repository.UserRepository;
 import in.koreatech.koin.domain.user.repository.UserTokenRepository;
 import in.koreatech.koin.global.auth.JwtProvider;
 import in.koreatech.koin.global.auth.exception.AuthorizationException;
-import in.koreatech.koin.global.domain.email.exception.DuplicationEmailException;
 import in.koreatech.koin.global.exception.KoinIllegalArgumentException;
 import lombok.RequiredArgsConstructor;
 
@@ -54,29 +50,24 @@ public class UserService {
 
     @Transactional
     public UserLoginResponse login(UserLoginRequest request) {
-        User user = userRepository.getByEmail(request.email());
-        Optional<StudentTemporaryStatus> studentTemporaryStatus = studentRedisRepository.findById(request.email());
-
-        if (!user.isSamePassword(passwordEncoder, request.password())) {
-            throw new KoinIllegalArgumentException("비밀번호가 틀렸습니다.");
-        }
-
-        if (studentTemporaryStatus.isPresent()) {
-            throw new AuthorizationException("미인증 상태입니다. 아우누리에서 인증메일을 확인해주세요");
-        }
+        User user = checkLoginCredentials(request.email(), request.password());
+        checkUserAuthentication(request.email());
 
         String accessToken = jwtProvider.createToken(user);
-        String refreshToken = String.format("%s-%d", UUID.randomUUID(), user.getId());
+        String refreshToken = generateRefreshToken(user);
         UserToken savedToken = userTokenRepository.save(UserToken.create(user.getId(), refreshToken));
-        user.updateLastLoggedTime(LocalDateTime.now());
-        User saved = userRepository.save(user);
 
-        return UserLoginResponse.of(accessToken, savedToken.getRefreshToken(), saved.getUserType().getValue());
+        updateLastLoginTime(user);
+        return UserLoginResponse.of(accessToken, savedToken.getRefreshToken(), user.getUserType().getValue());
     }
 
     @Transactional
     public void logout(Integer userId) {
         userTokenRepository.deleteById(userId);
+    }
+
+    public void checkLogin(String accessToken) {
+        jwtProvider.getUserId(accessToken);
     }
 
     public UserTokenRefreshResponse refresh(UserTokenRefreshRequest request) {
@@ -89,14 +80,6 @@ public class UserService {
 
         String accessToken = jwtProvider.createToken(user);
         return UserTokenRefreshResponse.of(accessToken, userToken.getRefreshToken());
-    }
-
-    private String getUserId(String refreshToken) {
-        String[] split = refreshToken.split("-");
-        if (split.length == 0) {
-            throw new AuthorizationException("올바르지 않은 인증 토큰입니다. refreshToken: " + refreshToken);
-        }
-        return split[split.length - 1];
     }
 
     @Transactional
@@ -112,29 +95,6 @@ public class UserService {
         eventPublisher.publishEvent(new UserDeleteEvent(user.getEmail(), user.getUserType()));
     }
 
-    public void checkLogin(String accessToken) {
-        jwtProvider.getUserId(accessToken);
-    }
-
-    public void checkPassword(UserPasswordCheckRequest request, Integer userId) {
-        User user = userRepository.getById(userId);
-        if (!user.isSamePassword(passwordEncoder, request.password())) {
-            throw new KoinIllegalArgumentException("올바르지 않은 비밀번호입니다.");
-        }
-    }
-
-    public void checkExistsEmail(EmailCheckExistsRequest request) {
-        userRepository.findByEmail(request.email()).ifPresent(user -> {
-            throw DuplicationEmailException.withDetail("email: " + user.getEmail());
-        });
-    }
-
-    public void checkUserNickname(NicknameCheckExistsRequest request) {
-        userRepository.findByNickname(request.nickname()).ifPresent(user -> {
-            throw DuplicationNicknameException.withDetail("nickname: " + request.nickname());
-        });
-    }
-
     public AuthResponse getAuth(Integer userId) {
         User user = userRepository.getById(userId);
         return AuthResponse.from(user);
@@ -143,5 +103,37 @@ public class UserService {
     public CoopResponse getCoop(Integer userId) {
         User user = userRepository.getById(userId);
         return CoopResponse.from(user);
+    }
+
+    private String getUserId(String refreshToken) {
+        String[] split = refreshToken.split("-");
+        if (split.length == 0) {
+            throw new AuthorizationException("올바르지 않은 인증 토큰입니다. refreshToken: " + refreshToken);
+        }
+        return split[split.length - 1];
+    }
+
+    public User checkLoginCredentials(String email, String password) {
+        User user = userRepository.getByEmail(email);
+        if (!user.isSamePassword(passwordEncoder, password)) {
+            throw new KoinIllegalArgumentException("비밀번호가 틀렸습니다.");
+        }
+        return user;
+    }
+
+    public void checkUserAuthentication(String email) {
+        Optional<StudentTemporaryStatus> studentTemporaryStatus = studentRedisRepository.findById(email);
+        if (studentTemporaryStatus.isPresent()) {
+            throw new AuthorizationException("미인증 상태입니다. 아우누리에서 인증메일을 확인해주세요");
+        }
+    }
+
+    public String generateRefreshToken(User user) {
+        return String.format("%s-%d", UUID.randomUUID(), user.getId());
+    }
+
+    public void updateLastLoginTime(User user) {
+        user.updateLastLoggedTime(LocalDateTime.now());
+        userRepository.save(user);
     }
 }
